targetDecomojiList を for で fetch する処理を importer.js モジュールに切り出した
